+++ pkg/deployment/resources/certificates_client_auth.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2016-2022 ArangoDB GmbH, Cologne, Germany
+// Copyright 2016-2024 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -23,36 +23,23 @@ package resources
 import (
 	"context"
 	"fmt"
-	"time"
 
 	meta "k8s.io/apimachinery/pkg/apis/meta/v1"
-
-	certificates "github.com/arangodb-helper/go-certificates"
 
 	api "github.com/arangodb/kube-arangodb/pkg/apis/deployment/v1"
 	"github.com/arangodb/kube-arangodb/pkg/util/errors"
 	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil"
 	secretv1 "github.com/arangodb/kube-arangodb/pkg/util/k8sutil/inspector/secret/v1"
 	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil/kerrors"
-)
-
-const (
-	clientAuthECDSACurve = "P256" // This curve is the default that ArangoDB accepts and plenty strong
+	ktls "github.com/arangodb/kube-arangodb/pkg/util/k8sutil/tls"
 )
 
 // createClientAuthCACertificate creates a client authentication CA certificate and stores it in a secret with name
 // specified in the given spec.
 func (r *Resources) createClientAuthCACertificate(ctx context.Context, secrets secretv1.ModInterface, spec api.SyncAuthenticationSpec, deploymentName string, ownerRef *meta.OwnerReference) error {
 	log := r.log.Str("section", "secrets")
-	options := certificates.CreateCertificateOptions{
-		CommonName:   fmt.Sprintf("%s Client Authentication Root Certificate", deploymentName),
-		ValidFrom:    time.Now(),
-		ValidFor:     caTTL,
-		IsCA:         true,
-		IsClientAuth: true,
-		ECDSACurve:   clientAuthECDSACurve,
-	}
-	cert, priv, err := certificates.CreateCertificate(options, nil)
+
+	cert, priv, err := ktls.CreateTLSCACertificate(fmt.Sprintf("%s Client Authentication Root Certificate", deploymentName))
 	if err != nil {
 		log.Err(err).Str("name", spec.GetClientCASecretName()).Debug("Failed to create CA certificate")
 		return errors.WithStack(err)