+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - (Feature) Custom Gateway image
 - (Bugfix) Fix race condition in ArangoBackup
 - (Feature) Improve Gateway Config gen
+- (Feature) Integration Service TLS
 
 ## [1.2.42](https://github.com/arangodb/kube-arangodb/tree/1.2.42) (2024-07-23)
 - (Maintenance) Go 1.22.4 & Kubernetes 1.29.6 libraries